requirements.txt
anthropic==0.18.1
flask==2.3.3
python-dotenv==1.0.0